done task

+++ handlers/admin_handlers.py
@@ -1,5 +1,5 @@
 from aiogram import types, Dispatcher
-
+from config_data.config import get_admins
 from lexicon import choise_random_gif, ANSWER_ADMIN
 
 
@@ -9,16 +9,15 @@ def register_admin_handler(dp: Dispatcher):
 
 # Commad Start from admins 
 async def send_command_start_admin(message: types.Message):
-    match message.chat.id or message.from_user.id:
-        case 222997056:
+    for admin in get_admins():
+        if admin == get_admins()[0]:
             await message.bot.send_animation(
                 chat_id=message.chat.id,
-                animation=choise_random_gif("hello"),                     
+                animation=choise_random_gif("hello"),
                 caption=ANSWER_ADMIN["Victor"])
             await message.delete()
-
-        case 1057974570:
-            await message.bot.send_animation(chat_id=message.chat.id,
-                animation=choise_random_gif("hello"),
-                caption=ANSWER_ADMIN["Kristina"])
-            await message.delete()
+        # если по другому тогда отправить другому админу
+        await message.bot.send_animation(chat_id=message.chat.id,
+                                         animation=choise_random_gif("hello"),
+                                         caption=ANSWER_ADMIN["Kristina"])
+        await message.delete()
